+++ egui_web/src/webgl1.rs
@@ -289,6 +289,21 @@ impl epi::NativeTexture for WebGlPainter {
 }
 
 impl crate::Painter for WebGlPainter {
+    fn max_texture_side(&self) -> usize {
+        if let Ok(max_texture_side) = self
+            .gl
+            .get_parameter(web_sys::WebGlRenderingContext::MAX_TEXTURE_SIZE)
+        {
+            if let Some(max_texture_side) = max_texture_side.as_f64() {
+                return max_texture_side as usize;
+            }
+        }
+
+        crate::console_error("Failed to query max texture size");
+
+        2048
+    }
+
     fn set_texture(&mut self, tex_id: egui::TextureId, delta: &egui::epaint::ImageDelta) {
         match &delta.image {
             egui::ImageData::Color(image) => {