+++ epaint/src/text/font.rs
@@ -1,6 +1,5 @@
 use crate::{
     mutex::{Arc, Mutex, RwLock},
-    text::TextStyle,
     TextureAtlas,
 };
 use ahash::AHashMap;
@@ -59,7 +58,7 @@ impl Default for GlyphInfo {
 pub struct FontImpl {
     ab_glyph_font: ab_glyph::FontArc,
     /// Maximum character height
-    scale_in_pixels: f32,
+    scale_in_pixels: u32,
     height_in_points: f32,
     // move each character by this much (hack)
     y_offset: f32,
@@ -73,20 +72,13 @@ impl FontImpl {
         atlas: Arc<Mutex<TextureAtlas>>,
         pixels_per_point: f32,
         ab_glyph_font: ab_glyph::FontArc,
-        scale_in_points: f32,
+        scale_in_pixels: u32,
         y_offset: f32,
     ) -> FontImpl {
-        assert!(scale_in_points > 0.0);
+        assert!(scale_in_pixels > 0);
         assert!(pixels_per_point > 0.0);
 
-        let scale_in_pixels = pixels_per_point * scale_in_points;
-
-        // Round to an even number of physical pixels to get even kerning.
-        // See https://github.com/emilk/egui/issues/382
-        let scale_in_pixels = scale_in_pixels.round();
-        let scale_in_points = scale_in_pixels / pixels_per_point;
-
-        let height_in_points = scale_in_points;
+        let height_in_points = scale_in_pixels as f32 / pixels_per_point;
 
         // TODO: use v_metrics for line spacing ?
         // let v = rusttype_font.v_metrics(Scale::uniform(scale_in_pixels));
@@ -162,7 +154,7 @@ impl FontImpl {
                 &mut self.atlas.lock(),
                 &self.ab_glyph_font,
                 glyph_id,
-                self.scale_in_pixels,
+                self.scale_in_pixels as f32,
                 self.y_offset,
                 self.pixels_per_point,
             );
@@ -180,7 +172,7 @@ impl FontImpl {
     ) -> f32 {
         use ab_glyph::{Font as _, ScaleFont};
         self.ab_glyph_font
-            .as_scaled(self.scale_in_pixels)
+            .as_scaled(self.scale_in_pixels as f32)
             .kern(last_glyph_id, glyph_id)
             / self.pixels_per_point
     }
@@ -202,23 +194,21 @@ type FontIndex = usize;
 // TODO: rename?
 /// Wrapper over multiple `FontImpl` (e.g. a primary + fallbacks for emojis)
 pub struct Font {
-    text_style: TextStyle,
     fonts: Vec<Arc<FontImpl>>,
     /// Lazily calculated.
-    characters: RwLock<Option<std::collections::BTreeSet<char>>>,
+    characters: Option<std::collections::BTreeSet<char>>,
     replacement_glyph: (FontIndex, GlyphInfo),
     pixels_per_point: f32,
     row_height: f32,
-    glyph_info_cache: RwLock<AHashMap<char, (FontIndex, GlyphInfo)>>,
+    glyph_info_cache: AHashMap<char, (FontIndex, GlyphInfo)>,
 }
 
 impl Font {
-    pub fn new(text_style: TextStyle, fonts: Vec<Arc<FontImpl>>) -> Self {
+    pub fn new(fonts: Vec<Arc<FontImpl>>) -> Self {
         if fonts.is_empty() {
             return Self {
-                text_style,
                 fonts,
-                characters: RwLock::new(None),
+                characters: None,
                 replacement_glyph: Default::default(),
                 pixels_per_point: 1.0,
                 row_height: 0.0,
@@ -230,9 +220,8 @@ impl Font {
         let row_height = fonts[0].row_height();
 
         let mut slf = Self {
-            text_style,
             fonts,
-            characters: RwLock::new(None),
+            characters: None,
             replacement_glyph: Default::default(),
             pixels_per_point,
             row_height,
@@ -260,26 +249,20 @@ impl Font {
             slf.glyph_info(c);
         }
         slf.glyph_info('°');
-        slf.glyph_info(crate::text::PASSWORD_REPLACEMENT_CHAR); // password replacement character
+        slf.glyph_info(crate::text::PASSWORD_REPLACEMENT_CHAR);
 
         slf
     }
 
     /// All supported characters
-    pub fn characters(&self) -> BTreeSet<char> {
-        if self.characters.read().is_none() {
+    pub fn characters(&mut self) -> &BTreeSet<char> {
+        self.characters.get_or_insert_with(|| {
             let mut characters = BTreeSet::new();
             for font in &self.fonts {
                 characters.extend(font.characters());
             }
-            self.characters.write().replace(characters);
-        }
-        self.characters.read().clone().unwrap()
-    }
-
-    #[inline(always)]
-    pub fn text_style(&self) -> TextStyle {
-        self.text_style
+            characters
+        })
     }
 
     #[inline(always)]
@@ -295,35 +278,30 @@ impl Font {
 
     pub fn uv_rect(&self, c: char) -> UvRect {
         self.glyph_info_cache
-            .read()
             .get(&c)
             .map(|gi| gi.1.uv_rect)
             .unwrap_or_default()
     }
 
     /// Width of this character in points.
-    pub fn glyph_width(&self, c: char) -> f32 {
+    pub fn glyph_width(&mut self, c: char) -> f32 {
         self.glyph_info(c).1.advance_width
     }
 
     /// `\n` will (intentionally) show up as the replacement character.
-    fn glyph_info(&self, c: char) -> (FontIndex, GlyphInfo) {
-        {
-            if let Some(font_index_glyph_info) = self.glyph_info_cache.read().get(&c) {
-                return *font_index_glyph_info;
-            }
+    fn glyph_info(&mut self, c: char) -> (FontIndex, GlyphInfo) {
+        if let Some(font_index_glyph_info) = self.glyph_info_cache.get(&c) {
+            return *font_index_glyph_info;
         }
 
         let font_index_glyph_info = self.glyph_info_no_cache_or_fallback(c);
         let font_index_glyph_info = font_index_glyph_info.unwrap_or(self.replacement_glyph);
-        self.glyph_info_cache
-            .write()
-            .insert(c, font_index_glyph_info);
+        self.glyph_info_cache.insert(c, font_index_glyph_info);
         font_index_glyph_info
     }
 
     #[inline]
-    pub(crate) fn glyph_info_and_font_impl(&self, c: char) -> (Option<&FontImpl>, GlyphInfo) {
+    pub(crate) fn glyph_info_and_font_impl(&mut self, c: char) -> (Option<&FontImpl>, GlyphInfo) {
         if self.fonts.is_empty() {
             return (None, self.replacement_glyph.1);
         }
@@ -332,12 +310,10 @@ impl Font {
         (Some(font_impl), glyph_info)
     }
 
-    fn glyph_info_no_cache_or_fallback(&self, c: char) -> Option<(FontIndex, GlyphInfo)> {
+    fn glyph_info_no_cache_or_fallback(&mut self, c: char) -> Option<(FontIndex, GlyphInfo)> {
         for (font_index, font_impl) in self.fonts.iter().enumerate() {
             if let Some(glyph_info) = font_impl.glyph_info(c) {
-                self.glyph_info_cache
-                    .write()
-                    .insert(c, (font_index, glyph_info));
+                self.glyph_info_cache.insert(c, (font_index, glyph_info));
                 return Some((font_index, glyph_info));
             }
         }
@@ -379,11 +355,11 @@ fn allocate_glyph(
                 if v > 0.0 {
                     let px = glyph_pos.0 + x as usize;
                     let py = glyph_pos.1 + y as usize;
-                    image[(px, py)] = (v * 255.0).round() as u8;
+                    image[(px, py)] = fast_round(v * 255.0);
                 }
             });
 
-            let offset_in_pixels = vec2(bb.min.x as f32, scale_in_pixels as f32 + bb.min.y as f32);
+            let offset_in_pixels = vec2(bb.min.x as f32, scale_in_pixels + bb.min.y as f32);
             let offset = offset_in_pixels / pixels_per_point + y_offset * Vec2::Y;
             UvRect {
                 offset,
@@ -406,4 +382,8 @@ fn allocate_glyph(
         advance_width: advance_width_in_points,
         uv_rect,
     }
+}
+
+fn fast_round(r: f32) -> u8 {
+    (r + 0.5).floor() as _ // rust does a saturating cast since 1.45
 }